Move LearningResourcePage and TopicArticlePage to new folder, ArticlePage
