+++ src/containers/App/SubjectMatterPage.jsx
@@ -10,8 +10,8 @@ import PropTypes from 'prop-types';
 import { Route, Switch, withRouter } from 'react-router-dom';
 import NotFoundPage from '../NotFoundPage/NotFoundPage';
 import PrivateRoute from '../PrivateRoute/PrivateRoute';
-import TopicArticlePage from '../TopicArticlePage/TopicArticlePage';
-import LearningResourcePage from '../LearningResourcePage/LearningResourcePage';
+import TopicArticlePage from '../ArticlePage/TopicArticlePage/TopicArticlePage';
+import LearningResourcePage from '../ArticlePage/LearningResourcePage/LearningResourcePage';
 import Footer from './components/Footer';
 import { LocationShape, HistoryShape } from '../../shapes';
