+++ src/containers/ArticlePage/LearningResourcePage/components/LearningResourceForm.jsx
@@ -11,39 +11,38 @@ import PropTypes from 'prop-types';
 import { injectT } from '@ndla/i18n';
 import isEmpty from 'lodash/fp/isEmpty';
 import { Formik, Form } from 'formik';
+
 import {
-  topicArticleContentToHTML,
-  topicArticleContentToEditorValue,
+  learningResourceContentToHTML,
+  learningResourceContentToEditorValue,
   editorValueToPlainText,
   plainTextToEditorValue,
-} from '../../../util/articleContentConverter';
-import { parseEmbedTag, createEmbedTag } from '../../../util/embedTagHelpers';
-import { LicensesArrayOf, ArticleShape } from '../../../shapes';
+} from '../../../../util/articleContentConverter';
+import { LicensesArrayOf, ArticleShape } from '../../../../shapes';
+import { FormikAlertModalWrapper, formClasses } from '../../../FormikForm';
+import validateFormik from '../../../../components/formikValidationSchema';
+import LearningResourcePanels from './LearningResourcePanels';
 import {
   DEFAULT_LICENSE,
   parseCopyrightContributors,
   isFormikFormDirty,
-  topicArticleRules,
   parseImageUrl,
-} from '../../../util/formHelper';
-import { FormikAlertModalWrapper, formClasses } from '../../FormikForm';
-import { toEditArticle } from '../../../util/routeHelpers';
-import { nullOrUndefined } from '../../../util/articleUtil';
-import validateFormik from '../../../components/formikValidationSchema';
-import TopicArticleAccordionPanels from './TopicArticleAccordionPanels';
-import HeaderWithLanguage from '../../../components/HeaderWithLanguage';
-import EditorFooter from '../../../components/SlateEditor/EditorFooter';
-import { useArticleFormHooks } from '../../FormikForm/articleFormHooks';
-import usePreventWindowUnload from '../../FormikForm/preventWindowUnloadHook';
-import Spinner from '../../../components/Spinner';
+  learningResourceRules,
+} from '../../../../util/formHelper';
+import { toEditArticle } from '../../../../util/routeHelpers';
+import { nullOrUndefined } from '../../../../util/articleUtil';
+import HeaderWithLanguage from '../../../../components/HeaderWithLanguage';
+import EditorFooter from '../../../../components/SlateEditor/EditorFooter';
+import { useArticleFormHooks } from '../../../FormikForm/articleFormHooks';
+import usePreventWindowUnload from '../../../FormikForm/preventWindowUnloadHook';
+import Spinner from '../../../../components/Spinner';
 
 export const getInitialValues = (article = {}) => {
-  const visualElement = parseEmbedTag(article.visualElement);
   const metaImageId = parseImageUrl(article.metaImage);
   return {
     agreementId: article.copyright ? article.copyright.agreementId : undefined,
-    articleType: 'topic-article',
-    content: topicArticleContentToEditorValue(article.content),
+    articleType: 'standard',
+    content: learningResourceContentToEditorValue(article.content),
     creators: parseCopyrightContributors(article, 'creators'),
     id: article.id,
     introduction: plainTextToEditorValue(article.introduction, true),
@@ -53,6 +52,7 @@ export const getInitialValues = (article = {}) => {
     metaImageAlt: article.metaImage?.alt || '',
     metaImageId,
     notes: [],
+    origin: article.copyright?.origin || '',
     processors: parseCopyrightContributors(article, 'processors'),
     published: article.published,
     revision: article.revision,
@@ -61,9 +61,8 @@ export const getInitialValues = (article = {}) => {
     supportedLanguages: article.supportedLanguages || [],
     tags: article.tags || [],
     slatetitle: plainTextToEditorValue(article.title, true),
-    updated: article.updated,
     updatePublished: false,
-    visualElement: visualElement || {},
+    updated: article.updated,
     grepCodes: article.grepCodes || [],
     conceptIds: article.conceptIds || [],
     availability: article.availability || 'everyone',
@@ -78,18 +77,17 @@ const getPublishedDate = (values, initialValues, preview = false) => {
     return values.published;
   }
 
-  const hasPublishedDateChaned = initialValues.published !== values.published;
-  if (hasPublishedDateChaned || values.updatePublished) {
+  const hasPublishedDateChanged = initialValues.published !== values.published;
+  if (hasPublishedDateChanged || values.updatePublished) {
     return values.published;
   }
   return undefined;
 };
 
-// TODO preview parameter does not work for topic articles. Used from PreviewDraftLightbox
-const getArticleFromSlate = ({ values, initialValues, licenses, preview = false }) => {
-  const emptyField = values.id ? '' : undefined;
-  const visualElement = createEmbedTag(isEmpty(values.visualElement) ? {} : values.visualElement);
-  const content = topicArticleContentToHTML(values.content);
+const getArticleFromSlate = ({ values, licenses, initialValues, preview = false }) => {
+  const content = learningResourceContentToHTML(values.content);
+  const emptyContent = values.id ? '' : undefined;
+
   const metaImage = values?.metaImageId
     ? {
         id: values.metaImageId,
@@ -98,26 +96,25 @@ const getArticleFromSlate = ({ values, initialValues, licenses, preview = false 
     : nullOrUndefined(values?.metaImageId);
 
   const article = {
-    articleType: 'topic-article',
-    content: content || emptyField,
+    articleType: 'standard',
+    content: content && content.length > 0 ? content : emptyContent,
     copyright: {
       license: licenses.find(license => license.license === values.license),
+      origin: values.origin,
       creators: values.creators,
       processors: values.processors,
       rightsholders: values.rightsholders,
-      agreementId: values.agreementId,
     },
     id: values.id,
     introduction: editorValueToPlainText(values.introduction),
-    metaDescription: editorValueToPlainText(values.metaDescription),
     language: values.language,
     metaImage,
+    metaDescription: editorValueToPlainText(values.metaDescription),
     notes: values.notes || [],
     published: getPublishedDate(values, initialValues, preview),
     supportedLanguages: values.supportedLanguages,
     tags: values.tags,
     title: editorValueToPlainText(values.slatetitle),
-    visualElement: visualElement,
     grepCodes: values.grepCodes,
     conceptIds: values.conceptIds,
     availability: values.availability,
@@ -126,12 +123,11 @@ const getArticleFromSlate = ({ values, initialValues, licenses, preview = false 
   return article;
 };
 
-const TopicArticleForm = props => {
+const LearningResourceForm = props => {
   const {
     savedToServer,
     formikRef,
     initialValues,
-    setResetModal,
     handleSubmit,
     fetchStatusStateMachine,
     validateDraft,
@@ -139,38 +135,24 @@ const TopicArticleForm = props => {
   } = useArticleFormHooks({ getInitialValues, getArticleFromSlate, ...props });
   const [translateOnContinue, setTranslateOnContinue] = useState(false);
 
-  const {
-    t,
-    article,
-    updateArticle,
-    translating,
-    translateArticle,
-    licenses,
-    isNewlyCreated,
-    ...rest
-  } = props;
-
   const FormikChild = formik => {
     // eslint doesn't allow this to be inlined when using hooks (in usePreventWindowUnload)
     const { values, dirty, isSubmitting, setValues, errors, touched, ...formikProps } = formik;
-
     const formIsDirty = isFormikFormDirty({
       values,
       initialValues,
       dirty,
     });
     usePreventWindowUnload(formIsDirty);
-    const getArticle = () => getArticleFromSlate({ values, initialValues, licenses });
+    const getArticle = preview => getArticleFromSlate({ values, initialValues, licenses, preview });
     return (
       <Form {...formClasses()}>
         <HeaderWithLanguage
           values={values}
           content={article}
-          getEntity={getArticle}
           editUrl={lang => toEditArticle(values.id, values.articleType, lang)}
+          getEntity={getArticle}
           formIsDirty={formIsDirty}
-          getInitialValues={getInitialValues}
-          setValues={setValues}
           isSubmitting={isSubmitting}
           translateArticle={translateArticle}
           setTranslateOnContinue={setTranslateOnContinue}
@@ -179,12 +161,12 @@ const TopicArticleForm = props => {
         {translating ? (
           <Spinner withWrapper />
         ) : (
-          <TopicArticleAccordionPanels
+          <LearningResourcePanels
             values={values}
             errors={errors}
-            updateNotes={updateArticle}
             article={article}
             touched={touched}
+            updateNotes={updateArticle}
             formIsDirty={formIsDirty}
             getInitialValues={getInitialValues}
             setValues={setValues}
@@ -193,19 +175,23 @@ const TopicArticleForm = props => {
             fetchSearchTags={fetchSearchTags}
             {...formikProps}
             {...rest}
-            handleSubmit={() => handleSubmit(formik)}
+            handleSubmit={() => {
+              handleSubmit(formik, isNewlyCreated);
+            }}
           />
         )}
+
         <EditorFooter
           showSimpleFooter={!article.id}
           isSubmitting={isSubmitting}
           formIsDirty={formIsDirty}
           savedToServer={savedToServer}
           getEntity={getArticle}
-          showReset={() => setResetModal(true)}
           errors={errors}
           values={values}
-          onSaveClick={saveAsNewVersion => handleSubmit(formik, saveAsNewVersion)}
+          onSaveClick={saveAsNewVersion => {
+            handleSubmit(formik, saveAsNewVersion);
+          }}
           entityStatus={article.status}
           getStateStatuses={fetchStatusStateMachine}
           validateEntity={validateDraft}
@@ -225,35 +211,52 @@ const TopicArticleForm = props => {
     );
   };
 
+  const {
+    t,
+    article,
+    updateArticle,
+    translating,
+    translateArticle,
+    licenses,
+    isNewlyCreated,
+    ...rest
+  } = props;
   return (
     <Formik
       enableReinitialize={translating}
-      validateOnMount
       initialValues={initialValues}
-      validateOnChange={false}
       innerRef={formikRef}
+      validateOnBlur={false}
+      validateOnMount
       onSubmit={() => ({})}
-      validate={values => validateFormik(values, topicArticleRules, t)}>
+      validate={values => validateFormik(values, learningResourceRules, t)}>
       {FormikChild}
     </Formik>
   );
 };
 
-TopicArticleForm.propTypes = {
+LearningResourceForm.propTypes = {
+  licenses: LicensesArrayOf,
   revision: PropTypes.number,
   updateArticle: PropTypes.func.isRequired,
   createMessage: PropTypes.func.isRequired,
-  applicationError: PropTypes.func.isRequired,
   articleStatus: PropTypes.shape({
     current: PropTypes.string,
     other: PropTypes.arrayOf(PropTypes.string),
   }),
   updateArticleAndStatus: PropTypes.func,
-  licenses: LicensesArrayOf,
+  taxonomy: PropTypes.shape({
+    resourceTypes: PropTypes.array,
+    filter: PropTypes.array,
+    topics: PropTypes.array,
+    loading: PropTypes.bool,
+  }),
+  userAccess: PropTypes.string,
   article: ArticleShape,
+  applicationError: PropTypes.func.isRequired,
   translating: PropTypes.bool,
   translateArticle: PropTypes.func,
   isNewlyCreated: PropTypes.bool,
 };
 
-export default injectT(TopicArticleForm);
+export default injectT(LearningResourceForm);
