+++ src/containers/ArticlePage/LearningResourcePage/components/LearningResourcePanels.jsx
@@ -4,11 +4,11 @@ import { injectT } from '@ndla/i18n';
 import Accordion, { AccordionWrapper, AccordionBar, AccordionPanel } from '@ndla/accordion';
 import LearningResourceTaxonomy from './LearningResourceTaxonomy';
 import LearningResourceContent from './LearningResourceContent';
-import FormikConcepts from '../../FormikForm/FormikConcepts';
-import { FormikCopyright, VersionAndNotesPanel, FormikMetadata } from '../../FormikForm';
-import { TAXONOMY_WRITE_SCOPE } from '../../../constants';
-import FormikGrepCodes from '../../FormikForm/FormikGrepCodes';
-import FormikField from '../../../components/FormikField';
+import FormikConcepts from '../../../FormikForm/FormikConcepts';
+import { FormikCopyright, VersionAndNotesPanel, FormikMetadata } from '../../../FormikForm';
+import { TAXONOMY_WRITE_SCOPE } from '../../../../constants';
+import FormikGrepCodes from '../../../FormikForm/FormikGrepCodes';
+import FormikField from '../../../../components/FormikField';
 
 const panels = [
   {
